refactor: rewrite with ECMAScript notation
